added src directory and move the directories of my own
