+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import App from 'next/app'
-import GlobalStyle from 'components/utils/GlobalStyle'
+import GlobalStyle from 'src/components/utils/GlobalStyle'
 
 class MyApp extends App {
   // Only uncomment this method if you have blocking data requirements for
